move LostConnectionMessageBox to bapsf_motion.gui.configure.message_boxes
